stuck with bundling worker and user eval code
bundling the worker code seems okay, but evaling the users code while
using modules is proving to be difficult. This article is pretty
interesting but I didn't have any luck with it.
https://2ality.com/2019/10/eval-via-import.html
no matter what I tried, even the simple examples I would get a
`Cannot find module 'data:text/javascript;base64` type error when trying
to use import() I'm not sure if that's because compatibility with
dynamic imports in the context of the worker. see the bottom of
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/import